Get mTLS running, start/stop/status of jobs working. Need to debug streaming, probably through the job manager lib.
tc6
