+++ conf/server/server.crt
@@ -1,22 +1,23 @@
 -----BEGIN CERTIFICATE-----
-MIIDrTCCApWgAwIBAgIUKSsWHn8PzPcigUer0Esnbakip9YwDQYJKoZIhvcNAQEL
+MIID5TCCAs2gAwIBAgIUKSsWHn8PzPcigUer0Esnbakip9gwDQYJKoZIhvcNAQEL
 BQAwazELMAkGA1UEBhMCVVMxDjAMBgNVBAgMBVN0YXRlMQ0wCwYDVQQHDARDaXR5
 MRUwEwYDVQQKDAxPcmdhbml6YXRpb24xEDAOBgNVBAsMB09yZ1VuaXQxFDASBgNV
-BAMMC2V4YW1wbGUuY29tMB4XDTI1MDEyMTE4NTUxNVoXDTI2MDEyMTE4NTUxNVow
+BAMMC2V4YW1wbGUuY29tMB4XDTI1MDEyMjE0MTkzMloXDTI2MDEyMjE0MTkzMlow
 cjELMAkGA1UEBhMCVVMxDjAMBgNVBAgMBVN0YXRlMQ0wCwYDVQQHDARDaXR5MRUw
 EwYDVQQKDAxPcmdhbml6YXRpb24xEDAOBgNVBAsMB09yZ1VuaXQxGzAZBgNVBAMM
 EnNlcnZlci5leGFtcGxlLmNvbTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoC
-ggEBALC5Y9smbYyrDkmKIy2kdl3OSjIP9h0B+TJy5VVtzXn0y4iAuSPSEsVHKgVe
-ilBLr+tjxJy2LK+Q9o2ByOiaeUlvpqy0TKBBuqELW2PFerL1z98Lf92Zgps6ku5J
-Y0V0+GfFqQ5eEqyGDoUwdvI/w8/po2/4t8VQzJj6FYJNwCws9fC9a1fdQwyt/oVj
-3i52nt2vIL0r17yeZNMAqEgwZsN02YlI3pkAG84UXj3EF1/kgdFYA1nXAG5h5JeE
-rgVtwmxi++e9qCneAGne7tucg0X8qlsgBPaFNckUdWMPzS0uwHRFvDh+7zvkJbVB
-9EmNU9FgJNiTVg9YGQyJsIbsNEUCAwEAAaNCMEAwHQYDVR0OBBYEFPiJGRwhJKe1
-P+b+7r1WXDgxPSszMB8GA1UdIwQYMBaAFMEPa/WAYmp946uSj8lZUaEVFRKYMA0G
-CSqGSIb3DQEBCwUAA4IBAQAQRNwH2QC+MKvtS/bjWQZAPeJxdA6ijC4rPEHCB2fa
-0b5D96q0op1Afzh9qXsVzHkREUwJzdqVS3m3KpIOCNJfEu4Jsi3xCmfPnjMSx1nf
-Izsh5sR30gMX/8bW+qt5PtmQrk9MPViyrkznMCF14RL+zH82ldIeW45wdYXfxvT0
-kGhCq2jixvOQm+Ub1fuPC6XTiLgnErFlUoHMaEzgzvTyO4MGj3LfPH8WDV3Z3XKk
-eRXW6yy0GpqfiCxkN2JWUXXqVkx7DFmrcf1IRAVcfHOFrrU80bedrHgVBzr1C8V9
-p2o/1cHx85oqv8pQPAqdyhY3HYgCzY5AJHSiC3zA5iZv
+ggEBAKWCnqfJl8sRAhrTKk3MJZfS2TNebdcGexNhDGyZg3izm+Lp9Fi5a5zftywO
+5E7B9izCNtqz/A9uS8l86zujdmybRUKjg62IrAUuvR67RskywHDXgczbE5skSm0L
+e6CzqP+Ycsh7P+kjzgFdk8x+ZFBD1BScR0Lb1tsUvBtEc0bvdB3kwEXq5BMANubD
+rtW+OVfLnfEH/6F6AvTuAAom03WmzX0g18x2xKqXgQPVMWN4rJM/cXMHEooRQ7vF
+Zz8N/V5alDdB7Uw62bVMbyJDL7iO58gKhC2w3NAX3YBaqh0Pprs338LuFY8s1GGH
+oDKuV6qXa3GawQD70Vr1h/h1rYECAwEAAaN6MHgwCwYDVR0PBAQDAgQwMBMGA1Ud
+JQQMMAoGCCsGAQUFBwMBMBQGA1UdEQQNMAuCCWxvY2FsaG9zdDAdBgNVHQ4EFgQU
+0bEq8HrUfX+KWDCn3RAlc0QIkq0wHwYDVR0jBBgwFoAUNrIknIOvjNNzF1mnLfbz
+WRPdFYswDQYJKoZIhvcNAQELBQADggEBABhoaehaAkDG6lHQrcs4csjfMNZX8bSm
+35H4tM/46B9A5vCCfVmgvwjLKrySvQEap/RQSPEFHgnuznM6XIPMSZq5nEQj0225
+lsKtmvkWeh2FbI1BKP7ksGnHo2roTXahIeCdIuK14PsjiDLRzISpBzOzXXAy8vFd
+Jx+CUtE2kPKeFRBiEgJjEztBvZIhZQTXxc90eSc3lz038QU3dRJ9EUY8pp00MlXk
+fQli19TuvgjxPjpYdFkmfisAVV3wR4TgKvXLEexW8p53PYz5mhU96db5SYUHYz4M
+HXI1kMNGlaJBok9IrlWYNuAbL8CLKK1OjeQwQZa/GQfw7QHb1h6Mzb8=
 -----END CERTIFICATE-----
