+++ lib/cli/stream_command.go
@@ -2,9 +2,9 @@ package cli
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/google/uuid"
+
 	"github.com/mcampo84/teleport_challenge/lib/client"
 )
 
@@ -21,16 +21,10 @@ func NewStreamCommand(client *client.Client, jobId uuid.UUID) *StreamCommand {
 }
 
 func (c *StreamCommand) Execute(ctx context.Context) (err error) {
-	err = c.client.StreamOutput(ctx, c.jobId, c)
+	err = c.client.StreamOutput(ctx, c.jobId)
 	if err != nil {
 		return err
 	}
-
-	return nil
-}
-
-func (c *StreamCommand) Receive(output []byte) error {
-	fmt.Println(output)
 
 	return nil
 }
