+++ lib/server/streaming_server_adapter.go
@@ -1,6 +1,8 @@
 package server
 
 import (
+	"fmt"
+
 	pb "github.com/mcampo84/teleport_challenge/lib/job_manager/pb/v1"
 )
 
@@ -15,6 +17,8 @@ func NewStreamingServerAdapter(streamingServer pb.CommandService_StreamOutputSer
 
 func (s *StreamingServerAdapter) Send(output []byte) error {
 	s.part++
+
+	fmt.Printf("Received %s as part %d\n", output, s.part)
 
 	return s.streamingServer.Send(&pb.StreamOutputResponse{
 		Part: s.part,
