+++ scripts/generate_certs.sh
@@ -14,8 +14,22 @@ openssl req -x509 -new -nodes -key "$SCRIPT_DIR/../conf/ca.key" -sha256 -days 36
 openssl genpkey -algorithm RSA -out "$SCRIPT_DIR/../conf/server/server.key"
 openssl req -new -key "$SCRIPT_DIR/../conf/server/server.key" -out "$SCRIPT_DIR/../conf/server/server.csr" -subj "/C=US/ST=State/L=City/O=Organization/OU=OrgUnit/CN=server.example.com"
 
+# Create a config file for the server certificate with SAN
+cat > "$SCRIPT_DIR/../conf/server/server_cert_config.cnf" <<EOL
+[req]
+distinguished_name = req_distinguished_name
+req_extensions = v3_req
+[req_distinguished_name]
+[v3_req]
+keyUsage = keyEncipherment, dataEncipherment
+extendedKeyUsage = serverAuth
+subjectAltName = @alt_names
+[alt_names]
+DNS.1 = localhost
+EOL
+
 # Sign server CSR with CA certificate
-openssl x509 -req -in "$SCRIPT_DIR/../conf/server/server.csr" -CA "$SCRIPT_DIR/../conf/ca.crt" -CAkey "$SCRIPT_DIR/../conf/ca.key" -CAcreateserial -out "$SCRIPT_DIR/../conf/server/server.crt" -days 365 -sha256
+openssl x509 -req -in "$SCRIPT_DIR/../conf/server/server.csr" -CA "$SCRIPT_DIR/../conf/ca.crt" -CAkey "$SCRIPT_DIR/../conf/ca.key" -CAcreateserial -out "$SCRIPT_DIR/../conf/server/server.crt" -days 365 -sha256 -extfile "$SCRIPT_DIR/../conf/server/server_cert_config.cnf" -extensions v3_req
 
 # Generate client key and CSR
 openssl genpkey -algorithm RSA -out "$SCRIPT_DIR/../conf/client/client.key"
